+++ google/CHANGELOG.md
@@ -4,7 +4,7 @@ This file documents changes to the `workbench-google` library, including notes o
 
 ## 0.21
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google" % "0.21-64a7b29"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google" % "0.21-24dcd92"`
 
 ### Added
 
@@ -14,6 +14,7 @@ SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google" % "0.
   for backwards compatibility.
 - `getProjectNumber` in `GoogleProjectDAO`
 - `addWorkloadIdentityUserRoleForUser` in `GoogleIamDAO`
+- Added enableRequesterPays for workspaces
 
 ### Changed